// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/lastikler/models/lastik-ekle.ts
export class LastikEkle {
    FirmaID: number;

    LastikMarkaID: number;

    LastikMarkaDesenID: number;

    SeriNo: string;

    Fiyat: number;

    EbatID: number;

    LastikTurID: number;

    LastikTipID: number;

    DisSeviyesi: number;

    LastikKilometre: number;

    LastikKonumID: number;
    
    KayitTarihi: Date

}